Don't use short module syntax, move file to correct place Issue #10575
