+++ spec/mailers/previews/devise.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# :nodoc:
+module Devise
+end
