--- spec/mailers/previews/devise_mailer_preview.rb
@@ -1,27 +1,0 @@
-# Preview all emails at http://localhost:3000/rails/mailers/devise_mailer
-class Devise::MailerPreview < ActionMailer::Preview
-
-  def confirmation_instructions
-    Devise::Mailer.confirmation_instructions(User.last, "faketoken")
-  end
-
-  def email_changed
-    Devise::Mailer.email_changed(User.last, User.last.email)
-  end
-
-  def invitation_instructions
-    Devise::Mailer.email_changed(User.last, "faketoken")
-  end
-
-  def password_changed
-    Devise::Mailer.email_changed(User.last)
-  end
-
-  def reset_password_instructions
-    Devise::Mailer.email_changed(User.last, "faketoken")
-  end
-
-  def unlock_instructions
-    Devise::Mailer.reset_password_instructions(User.last, "faketoken")
-  end
-end
